feat: include all possible letters of a given language
When the "Letters Only" option is enabled in the custom text mode,
include all possible letters. Previously we included only letters
from a language alphabet, but the alphabet usually includes only
a small set of letters. It allowed a situation in which a letter was
available on a keyboard, but still was excluded from lessons.

closes #212