refactor: do not generate `*.less.d.ts` files
It is a bit painful to keep these files in sync. However, life is not
bad without them.

Cons: TypeScript will compile anything, even if an imported class name
is missing. The bundler, however, will fail. So the error discovery is
delayed.

Pros: WebStorm is perfectly able to auto-complete the imported class
names even without the generated type definitions.