feat: do not round up the percent values to 100%
When there are not enough decimal places, a percent value can become
rounded up to 100%, event if it's less than that. This creates a false
impression on the users.

Instead we add as many decimal places as needed to such values.

So if a value is not rounded to 100%, we honor the configured number
of decimal places.

And if a value would be rounded up to 100%, we add decimal places
to display it as `99.9%`, `99.99%`, `99.999%` and so on.