feat: start using the Node.js test runner and the Chai.js assertions
For a long time I wanted to get rid of the AVA test runner.

* AVA does not allow nested tests and test suites.
* AVA has a very minimal set of assertion functions.
* AVA runs tests concurrently and it causes too many troubles.

Now that the Node.js test runner has finally matured in version 22,
it's about time to start using it.

After me migrated the project to EcmaScript modules, it became possible
to import the Chai.js 5.0 assertion library. It has a really nice API
and it's a pleasure to work with it.

Unlike other libraries, Chai.js does not pollute the global namespace,
has no `magic` functionality, feels lightweight.