chore: extract package `color`
It was difficult to import this class from the package `widgets` because
along with the TypeScript source code it also dragged component CSS
styles with it.

This is not a problem for the bundler (like webpack), but it was
difficult to use this class in 'normal' packages, like CLI tools.