fix the failing tests by disabling the asset loading
The tests were failing because the sound settings component was loading
sound assets asynchronously through a web request, and we did not fake
the response.

I think this is just a workaround, and a proper fix must be implemented.
I don't understand which change broke the tests, need to dig deper.